requirements.txt
asgiref==3.5.2
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
dlib==19.24.0
drf-extra-fields==3.4.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.22.4
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
XlsxWriter==3.0.3
